+++ tests/test_intrange.py
@@ -1,0 +1,6 @@
+import pytest
+
+from spans import intrange
+
+def test_len():
+    assert len(intrange(0, 5)) == 5